+++ src/main/java/com/example/kakaoshop/user/User.java
@@ -15,7 +15,7 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     @Column(length = 100, nullable = false, unique = true)
     private String email; // 인증시 필요한 필드
     @Column(length = 256, nullable = false)
@@ -27,7 +27,7 @@ public class User {
     private String roles; // role은 한 개 이상
 
     @Builder
-    public User(int id, String email, String password, String username, String roles) {
+    public User(Long id, String email, String password, String username, String roles) {
         this.id = id;
         this.email = email;
         this.password = password;
